Moved types to Math namespace
