+++ Src/Newtonsoft.Json.UnityConverters.Tests/ConvertingUnityTypes/Math/ColorTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes
+namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.Math
 {
     public class ColorTests : ValueTypeTester<Color>
     {
