+++ Src/Newtonsoft.Json.UnityConverters.Tests/ConvertingUnityTypes/Math/Matrix4x4Tests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes
+namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.Math
 {
     public class Matrix4x4Tests : ValueTypeTester<Matrix4x4>
     {
@@ -11,7 +11,7 @@ namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes
                 new Vector4(01f, 11f, 21f, 31f),
                 new Vector4(02f, 12f, 22f, 32f),
                 new Vector4(03f, 13f, 23f, 33f)
-            ), new { 
+            ), new {
                 m00 = 00f, m01 = 01f, m02 = 02f, m03 = 03f,
                 m10 = 10f, m11 = 11f, m12 = 12f, m13 = 13f,
                 m20 = 20f, m21 = 21f, m22 = 22f, m23 = 23f,
