+++ Src/Newtonsoft.Json.UnityConverters.Tests/ConvertingUnityTypes/Math/VectorTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes
+namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.Math
 {
     public class Vector2Tests : ValueTypeTester<Vector2>
     {
